dapp-refactor: moved erc/hooks to main hooks folder and erc/methods to main methods folder
