+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-20/methods/balance-of/index.tsx
@@ -25,8 +25,8 @@ import { balanceOf } from '@/api/hedera/erc20-interactions';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import HederaCommonTextField from '@/components/common/components/HederaCommonTextField';
 import { useToast, TableContainer, Table, Thead, Tr, Th, Tbody } from '@chakra-ui/react';
-import useUpdateMapStateUILocalStorage from '../../../shared/hooks/useUpdateMapStateUILocalStorage';
-import useRetrieveMapValueFromLocalStorage from '../../../shared/hooks/useRetrieveMapValueFromLocalStorage';
+import useUpdateMapStateUILocalStorage from '../../../../../../hooks/useUpdateMapStateUILocalStorage';
+import useRetrieveMapValueFromLocalStorage from '../../../../../../hooks/useRetrieveMapValueFromLocalStorage';
 import {
   HEDERA_BRANDING_COLORS,
   HEDERA_CHAKRA_TABLE_VARIANTS,
