+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-20/methods/token-permission/index.tsx
@@ -27,7 +27,7 @@ import { IoRefreshOutline } from 'react-icons/io5';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
-import { copyContentToClipboard } from '../../../shared/methods/common';
+import { copyContentToClipboard } from '../../../../../common/methods/common';
 import { getArrayTypedValuesFromLocalStorage } from '@/api/localStorage';
 import MultiLineMethod from '@/components/common/components/MultiLineMethod';
 import { handleErc20TokenPermissions } from '@/api/hedera/erc20-interactions';
