+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-721/methods/approve/index.tsx
@@ -32,8 +32,8 @@ import HederaCommonTextField from '@/components/common/components/HederaCommonTe
 import { Th, Tr, Table, Tbody, Thead, useToast, TableContainer } from '@chakra-ui/react';
 import { useUpdateTransactionResultsToLocalStorage } from '@/hooks/useUpdateLocalStorage';
 import { approveERC721ParamFields } from '@/utils/contract-interactions/erc/erc721/constant';
-import useUpdateMapStateUILocalStorage from '../../../shared/hooks/useUpdateMapStateUILocalStorage';
-import useRetrieveMapValueFromLocalStorage from '../../../shared/hooks/useRetrieveMapValueFromLocalStorage';
+import useUpdateMapStateUILocalStorage from '../../../../../../hooks/useUpdateMapStateUILocalStorage';
+import useRetrieveMapValueFromLocalStorage from '../../../../../../hooks/useRetrieveMapValueFromLocalStorage';
 import { handleRetrievingTransactionResultsFromLocalStorage } from '@/components/common/methods/handleRetrievingTransactionResultsFromLocalStorage';
 import {
   CONTRACT_NAMES,
